% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/M2B1.meta.R
\name{M2B1.meta}
\alias{M2B1.meta}
\title{Apply mediation meta analysis for binary outcome with one binary mediator}
\usage{
M2B1.meta(
  data,
  b0.ZM = 0,
  b0.ZMY = 0,
  cov.n = 0,
  pred.val = NULL,
  int = F,
  scale,
  boots.n = 1000,
  parallel = F
)
}
\arguments{
\item{data}{name of dataset. With matching data, the order of data columns: matching no, exposure, mediator, outcome, covariates. With non-matching data, the order of data columns: exposure, mediator, outcome, covariates}

\item{b0.ZM}{intercept of modeling the mediator as a dependent variable, and exposure as an independent variable (for conditional logistic regression)}

\item{b0.ZMY}{intercept of modeling the outcome as a dependent variable, exposure, and mediator as independent variables (for conditional logistic regression)}

\item{cov.n}{number of covariates}

\item{pred.val}{covariates values (default is median)}

\item{int}{interaction of exposure and mediator, TRUE or FALSE}

\item{scale}{arithmetic scale. "RD" for risk difference, "RR" for risk ratio, "OR" for odds ratio}

\item{boots.n}{number of bootstrapping}

\item{parallel}{parallel computing, TRUE or FALSE}
}
\description{
function: M2B1 for mediation analysis, M2B1.meta for mediation meta analysis
}
\examples{
# matching data
M2B1.meta(data=data,b0.ZM=0.01,b0.ZMY=0.02,cov.n=2,pred.val=c(1,35),int=T,scale="OR",boots.n=100,parallel=T)
# non-matching data
M2B1.meta(data=data,cov.n=2,pred.val=c(1,35),int=T,scale="OR",boots.n=100,parallel=F)
}
\keyword{Mediation}
\keyword{analysis}
\keyword{analysis,}
\keyword{meta}
